requirements.txt
pytest==8.3.3
pytest-cov==6.0.0
allure-pytest==2.13.5
allure-python-commons==2.13.5
selenium==4.28.1
selenium-wire==5.1.0
webdriver-manager==4.0.2